Avoid SDRAM ring buffer deadlock
Indicate wish to terminate the stream in BLOCKED state to avoid deadlock
when the SDRAM ring buffer is empty and SDRAM Host Read is bus master.
The only way out of BLOCKED state is for SDRAM Sink to write something
to the SDRAM ring buffer and that requires SDRAM Sink to be bus master.

Do not advance the read pointer unconditionally on wrap to avoid double
ring buffer read. It should be noted that I did not empirically manage
to get the double ring buffer read but avoiding the condition completely
is much easier than trying to prove that it is impossible to happen.

The SDRAM deadlock was empirically observed under heavy load with:
  * unconditional read and write pointer advancement on wrap, or
  * conditional (not blocked) read and write pointer advancement on wrap